+++ Changelog.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.26.1] - 2021-05-01
+## Fixed
+- Fix missing dependencies avaibable on `google()` repository
+
+## [0.26.0] - 2021-05-01
 ## Fixed
 - Fix lazy Test task configuration
 
@@ -209,7 +215,8 @@ The new feature is available as automatically registered gradle task, but also c
 [Unreleased]: https://github.com/usefulness/project-starter/compare/release/0.23.0...HEAD
 [0.29.0]: https://github.com/usefulness/project-starter/compare/release/0.28.0.../release/0.29.0
 [0.28.0]: https://github.com/usefulness/project-starter/compare/release/0.27.0.../release/0.28.0
-[0.27.0]: https://github.com/usefulness/project-starter/compare/release/0.26.0.../release/0.27.0
+[0.27.0]: https://github.com/usefulness/project-starter/compare/release/0.26.1.../release/0.27.0
+[0.26.1]: https://github.com/usefulness/project-starter/compare/release/0.26.0.../release/0.26.1
 [0.26.0]: https://github.com/usefulness/project-starter/compare/release/0.25.0.../release/0.26.0
 [0.25.0]: https://github.com/usefulness/project-starter/compare/release/0.24.0.../release/0.25.0
 [0.24.0]: https://github.com/usefulness/project-starter/compare/release/0.23.0.../release/0.24.0